refactor(#14): MemberRepository 패키지 변경
