+++ src/main/java/org/example/stellog/member/api/dto/response/MemberListInfoDto.java
@@ -1,0 +1,8 @@
+package org.example.stellog.member.api.dto.response;
+
+import java.util.List;
+
+public record MemberListInfoDto(
+        List<MemberInfoDto> memberInfoDtos
+) {
+}
