+++ src/main/java/org/example/stellog/member/domain/repository/MemberRepository.java
@@ -1,9 +1,13 @@
-package org.example.stellog.member.repository;
+package org.example.stellog.member.domain.repository;
 
-import java.util.Optional;
 import org.example.stellog.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+import java.util.Optional;
+
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByEmail(String email);
+
+    List<Member> findByNameContaining(String name);
 }
