+++ src/main/java/org/example/stellog/room/application/RoomService.java
@@ -3,8 +3,8 @@ package org.example.stellog.room.application;
 import lombok.RequiredArgsConstructor;
 import org.example.stellog.global.util.MemberRoomService;
 import org.example.stellog.member.domain.Member;
+import org.example.stellog.member.domain.repository.MemberRepository;
 import org.example.stellog.member.exception.MemberNotFoundException;
-import org.example.stellog.member.repository.MemberRepository;
 import org.example.stellog.room.api.dto.request.RoomRequestDto;
 import org.example.stellog.room.api.dto.response.MemberInfoDto;
 import org.example.stellog.room.api.dto.response.RoomDetailResponseDto;
